Add initial last polled entry.
This is a one off change that will be overwritten by the entity event
generator. The item changes will be ignored by terraform.

This allows us to keep the dynamo client as it currently is where there
is an error if the key is missing.

I've also increased the memory as 256Mb is a bit low for Java lambdas
and put it into the VPC so we can add the IP for the nat instance to the
Preservica security group/hosted preservica IP list.