Add eventbridge rules.
There are three rules.

One for the state change on our cloudwatch alarms when it moves to ALARM
One for the state change on our cloudwatch alarms when it moves to OK

The reason we need to is so we can have different slack icons for each.

Another one for sending messages directly to Slack.

I've removed the existing notifications and updated the queues to
remove the sns topic variable.
I've also deleted the notifications lambda.