Add more retries and add reserved concurrency
If we send more than one judgment message through at once, there is a
race condition on the upsert lambda so multiple of the same folder are
being created.

To prevent this, this sets the reserved concurrency to 1 and ups the
retries in the step function.

This is a temporary fix as it won't work if we have to migrate thousands
but it'll work for our testing.