Add mounted volume
There are errors in trusted advisor around the root file system being
writeable.
I've made two new volumes, one for /tests/boot and one for /temp.
These are for the tests to run in. The code which creates the docx and
sbt itself need a temp directory so they run in the /test/boot volume.

I also can't convince sbt to not use the /tmp directory so I've mounted
one for that as well.