DR2-1391 Increase S3 copy lambda memory
Sometimes, the s3 copy lambda is running out of memory.
As this is hopefully a temporary lambda, I don’t want to
spend time working out why so we just need to increase it.

I’ve tried it at 1024Mb and that works.

I've also updated the da-terraform-configurations commit hash.