DR2-1225 Fix KMS policies and bucket name
The bucket name had to be changed because sadly, prod-disaster-recovery
is taken.

The dev kms key was still using the old service_names parameter which
didn't do anything. I've updated it to use the service_details variable
which does.

I've also added the cloudwatch logs service to the main key. We're not
currently using it but we probably will.