DR2-1554 Add dr2 to names
So that we don't have to delete and recreate queues and step functions
and potentially miss messages. We're going to do this in two stages.
This is the first stage where we duplicate the queues, lambdas and step
function.
The latest code will then be deployed to the new lambda functions.
Then we'll disable the trigger on the parsed court document lambda and
move the subscription to the new topic.
Finally, the old queues, lambdas and step functions can be deleted.