Fixes for version check lambda.
The secret arn in the policy was always the live Preservica secret and
the secret in the environment variables was always the demo secret. I've
changed this so they match based on the environment.

The entry in the table for the version was a string and the scala code
was expecting it to be a number. I've updated it to be a number.