Increase court document handler memory
We had a failed judgment on prod because the lambda ran out of memory.
I suspect it was because there was a 12Mb log file in there and we have
to parse everything in the tar.gz file.

We could look into why this happened and try to fix the code but I don't
want to dedicate time to dealing with tar.gz files so I'll apply 21st
century solutions and throw more memory at it.