Step function updates
The distributed map which creates the asset opex needs to be batched otherwise we run out of lambda invocations.

The reconciler needs to stop outputing every failed child to Slack as this won’t work if there’s more than a few of them.

The reconciler also needs to send its output to S3 as it’s too large at the moment.

There are a few permissions updates needed and some of the SFN policy is out of date.