Step function fixes
The first is to suffix the StatePayload parameter so that it sends the
payload and not just a dollar.

The second is to stop sfn complaining about permissions. I'm not 100%
sure this is the mimimum needed but I mostly am.

The third is to selct out the StatePayload object from the upsert lamdba
input as it isn't expecting another top level object.